+++ driver/web/apis_tps.go
@@ -21,9 +21,11 @@ import (
 	"core-building-block/utils"
 	"encoding/json"
 	"net/http"
+	"strconv"
 	"strings"
 
 	"github.com/gorilla/mux"
+	"github.com/rokwire/core-auth-library-go/v2/authorization"
 	"github.com/rokwire/core-auth-library-go/v2/sigauth"
 	"github.com/rokwire/core-auth-library-go/v2/tokenauth"
 	"github.com/rokwire/logging-library-go/errors"
@@ -33,7 +35,8 @@ import (
 
 // TPSApisHandler handles the APIs implementation used by third-party services
 type TPSApisHandler struct {
-	coreAPIs *core.APIs
+	coreAPIs  *core.APIs
+	serviceID string
 }
 
 func (h TPSApisHandler) getServiceRegistrations(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
@@ -167,7 +170,126 @@ func (h TPSApisHandler) getServiceAccessTokens(l *logs.Log, r *http.Request, cla
 	return l.HttpResponseSuccessJSON(respData)
 }
 
+func (h TPSApisHandler) getAccounts(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
+	// get scopes relevant to accounts
+	if claims.Scope == "" {
+		return l.HttpResponseErrorData(logutils.StatusInvalid, model.TypeScope, nil, nil, http.StatusForbidden, false)
+	}
+
+	// appID and orgID
+	appID := r.URL.Query().Get("app_id")
+	orgID := r.URL.Query().Get("org_id")
+	if appID != "" && orgID != "" {
+		if !claims.AppOrg().CanAccessAppOrg(appID, orgID) {
+			return l.HttpResponseErrorData(logutils.StatusInvalid, model.TypeAppOrgPair, nil, nil, http.StatusForbidden, false)
+		}
+	} else {
+		appID = claims.AppID
+		orgID = claims.OrgID
+	}
+
+	// limit and offset
+	limit := 100
+	limitArg := r.URL.Query().Get("limit")
+	var err error
+	if limitArg != "" {
+		limit, err = strconv.Atoi(limitArg)
+		if err != nil {
+			return l.HttpResponseErrorAction(logutils.ActionParse, logutils.TypeArg, logutils.StringArgs("limit"), err, http.StatusBadRequest, false)
+		}
+	}
+	offset := 0
+	offsetArg := r.URL.Query().Get("offset")
+	if offsetArg != "" {
+		offset, err = strconv.Atoi(offsetArg)
+		if err != nil {
+			return l.HttpResponseErrorAction(logutils.ActionParse, logutils.TypeArg, logutils.StringArgs("offset"), err, http.StatusBadRequest, false)
+		}
+	}
+
+	var queryParams map[string]interface{}
+	err = json.NewDecoder(r.Body).Decode(&queryParams)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
+	}
+
+	// limit search params by scopes
+	scopes := claims.Scopes()
+	minAllAccessScope := authorization.Scope{ServiceID: h.serviceID, Resource: string(model.TypeAccount), Operation: authorization.ScopeOperationGet}
+	searchKeys := make([]string, 0)
+	for k := range queryParams {
+		searchKeys = append(searchKeys, k)
+	}
+	allAccess, approvedKeys, err := authorization.ResourceAccessForScopes(scopes, minAllAccessScope, searchKeys)
+	if err != nil {
+		return l.HttpResponseErrorData(logutils.StatusInvalid, "accounts query", nil, err, http.StatusForbidden, true)
+	}
+
+	accounts, err := h.coreAPIs.TPS.TPSGetAccounts(queryParams, appID, orgID, limit, offset, allAccess, approvedKeys)
+	if err != nil {
+		errFields := logutils.FieldArgs(queryParams)
+		return l.HttpResponseErrorAction(logutils.ActionGet, model.TypeAccount, &errFields, err, http.StatusInternalServerError, false)
+	}
+
+	respData, err := json.Marshal(accounts)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionMarshal, logutils.MessageDataType("accounts response"), nil, err, http.StatusInternalServerError, false)
+	}
+
+	return l.HttpResponseSuccessJSON(respData)
+}
+
 // NewTPSApisHandler creates new tps Handler instance
-func NewTPSApisHandler(coreAPIs *core.APIs) TPSApisHandler {
-	return TPSApisHandler{coreAPIs: coreAPIs}
+func NewTPSApisHandler(coreAPIs *core.APIs, serviceID string) TPSApisHandler {
+	return TPSApisHandler{coreAPIs: coreAPIs, serviceID: serviceID}
+}
+
+func (h TPSApisHandler) getAccountsCount(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
+	// get scopes relevant to accounts
+	if claims.Scope == "" {
+		return l.HttpResponseErrorData(logutils.StatusInvalid, model.TypeScope, nil, nil, http.StatusForbidden, false)
+	}
+
+	// appID and orgID
+	appID := r.URL.Query().Get("app_id")
+	orgID := r.URL.Query().Get("org_id")
+	if appID != "" && orgID != "" {
+		if !claims.AppOrg().CanAccessAppOrg(appID, orgID) {
+			return l.HttpResponseErrorData(logutils.StatusInvalid, model.TypeAppOrgPair, nil, nil, http.StatusForbidden, false)
+		}
+	} else {
+		appID = claims.AppID
+		orgID = claims.OrgID
+	}
+
+	var queryParams map[string]interface{}
+	err := json.NewDecoder(r.Body).Decode(&queryParams)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
+	}
+
+	// limit search params by scopes
+	scopes := claims.Scopes()
+	minAllAccessScope := authorization.Scope{ServiceID: h.serviceID, Resource: string(model.TypeAccount), Operation: authorization.ScopeOperationGet}
+	searchKeys := make([]string, 0)
+	for k := range queryParams {
+		searchKeys = append(searchKeys, k)
+	}
+	_, _, err = authorization.ResourceAccessForScopes(scopes, minAllAccessScope, searchKeys)
+	if err != nil {
+		return l.HttpResponseErrorData(logutils.StatusInvalid, "accounts count query", nil, err, http.StatusForbidden, true)
+	}
+
+	accounts, err := h.coreAPIs.TPS.TPsGetAccountsCount(queryParams, appID, orgID)
+	if err != nil {
+		errFields := logutils.FieldArgs(queryParams)
+		return l.HttpResponseErrorAction(logutils.ActionGet, model.TypeAccount, &errFields, err, http.StatusInternalServerError, false)
+	}
+
+	respData, err := json.Marshal(accounts)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionMarshal, logutils.MessageDataType("accounts count response"), nil, err, http.StatusInternalServerError, false)
+	}
+
+	return l.HttpResponseSuccessJSON(respData)
 }