+++ driver/web/docs/gen/gen_types.go
@@ -549,17 +549,18 @@ type Profile struct {
 
 // ProfileFields defines model for ProfileFields.
 type ProfileFields struct {
-	Address   *string `json:"address"`
-	BirthYear *int    `json:"birth_year"`
-	Country   *string `json:"country"`
-	Email     *string `json:"email"`
-	FirstName *string `json:"first_name,omitempty"`
-	Id        *string `json:"id,omitempty"`
-	LastName  *string `json:"last_name,omitempty"`
-	Phone     *string `json:"phone"`
-	PhotoUrl  *string `json:"photo_url,omitempty"`
-	State     *string `json:"state"`
-	ZipCode   *string `json:"zip_code"`
+	Address                *string                 `json:"address"`
+	BirthYear              *int                    `json:"birth_year"`
+	Country                *string                 `json:"country"`
+	Email                  *string                 `json:"email"`
+	FirstName              *string                 `json:"first_name,omitempty"`
+	Id                     *string                 `json:"id,omitempty"`
+	LastName               *string                 `json:"last_name,omitempty"`
+	Phone                  *string                 `json:"phone"`
+	PhotoUrl               *string                 `json:"photo_url,omitempty"`
+	State                  *string                 `json:"state"`
+	UnstructuredProperties *map[string]interface{} `json:"unstructured_properties"`
+	ZipCode                *string                 `json:"zip_code"`
 }
 
 // PubKey defines model for PubKey.
@@ -577,6 +578,7 @@ type ServiceAccount struct {
 	Name        string                      `json:"name"`
 	OrgId       string                      `json:"org_id"`
 	Permissions []string                    `json:"permissions"`
+	Scopes      []string                    `json:"scopes"`
 }
 
 // ServiceAccountCredential defines model for ServiceAccountCredential.
@@ -998,30 +1000,32 @@ type SharedReqParamsSetEmailCredential struct {
 
 // SharedReqProfile defines model for _shared_req_Profile.
 type SharedReqProfile struct {
-	Address   *string `json:"address"`
-	BirthYear *int    `json:"birth_year"`
-	Country   *string `json:"country"`
-	Email     *string `json:"email"`
-	FirstName *string `json:"first_name"`
-	LastName  *string `json:"last_name"`
-	Phone     *string `json:"phone"`
-	PhotoUrl  *string `json:"photo_url"`
-	State     *string `json:"state"`
-	ZipCode   *string `json:"zip_code"`
+	Address                *string                 `json:"address"`
+	BirthYear              *int                    `json:"birth_year"`
+	Country                *string                 `json:"country"`
+	Email                  *string                 `json:"email"`
+	FirstName              *string                 `json:"first_name"`
+	LastName               *string                 `json:"last_name"`
+	Phone                  *string                 `json:"phone"`
+	PhotoUrl               *string                 `json:"photo_url"`
+	State                  *string                 `json:"state"`
+	UnstructuredProperties *map[string]interface{} `json:"unstructured_properties"`
+	ZipCode                *string                 `json:"zip_code"`
 }
 
 // SharedReqProfileNullable defines model for _shared_req_ProfileNullable.
 type SharedReqProfileNullable struct {
-	Address   *string `json:"address"`
-	BirthYear *int    `json:"birth_year"`
-	Country   *string `json:"country"`
-	Email     *string `json:"email"`
-	FirstName *string `json:"first_name"`
-	LastName  *string `json:"last_name"`
-	Phone     *string `json:"phone"`
-	PhotoUrl  *string `json:"photo_url"`
-	State     *string `json:"state"`
-	ZipCode   *string `json:"zip_code"`
+	Address                *string                 `json:"address"`
+	BirthYear              *int                    `json:"birth_year"`
+	Country                *string                 `json:"country"`
+	Email                  *string                 `json:"email"`
+	FirstName              *string                 `json:"first_name"`
+	LastName               *string                 `json:"last_name"`
+	Phone                  *string                 `json:"phone"`
+	PhotoUrl               *string                 `json:"photo_url"`
+	State                  *string                 `json:"state"`
+	UnstructuredProperties *map[string]interface{} `json:"unstructured_properties"`
+	ZipCode                *string                 `json:"zip_code"`
 }
 
 // SharedReqRefresh defines model for _shared_req_Refresh.
@@ -1183,7 +1187,8 @@ type SystemReqCreateServiceAccount struct {
 	FirstParty  *bool                       `json:"first_party,omitempty"`
 	Name        *string                     `json:"name,omitempty"`
 	OrgId       string                      `json:"org_id"`
-	Permissions *[]string                   `json:"permissions,omitempty"`
+	Permissions *[]string                   `json:"permissions"`
+	Scopes      *[]string                   `json:"scopes"`
 }
 
 // SystemReqPermissions defines model for _system_req_permissions.
@@ -1224,8 +1229,9 @@ type SystemReqUpdateAuthType_Params struct {
 
 // SystemReqUpdateServiceAccount defines model for _system_req_update_service-account.
 type SystemReqUpdateServiceAccount struct {
-	Name        string   `json:"name"`
-	Permissions []string `json:"permissions"`
+	Name        *string   `json:"name,omitempty"`
+	Permissions *[]string `json:"permissions"`
+	Scopes      *[]string `json:"scopes"`
 }
 
 // DeleteAdminAccountMfaParams defines parameters for DeleteAdminAccountMfa.
@@ -1380,6 +1386,36 @@ type PostBbsAccessTokenJSONBody = ServicesReqServiceAccountsAccessToken
 
 // PostBbsAccessTokensJSONBody defines parameters for PostBbsAccessTokens.
 type PostBbsAccessTokensJSONBody = ServicesReqServiceAccountsAccessTokens
+
+// PostBbsAccountsJSONBody defines parameters for PostBbsAccounts.
+type PostBbsAccountsJSONBody = map[string]interface{}
+
+// PostBbsAccountsParams defines parameters for PostBbsAccounts.
+type PostBbsAccountsParams struct {
+	// The application ID to use to filter accounts
+	AppId *string `form:"app_id,omitempty" json:"app_id,omitempty"`
+
+	// The organization ID to use to filter accounts
+	OrgId *string `form:"org_id,omitempty" json:"org_id,omitempty"`
+
+	// The maximum number of accounts to return
+	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
+
+	// The index of the first account to return
+	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
+}
+
+// PostBbsAccountsCountJSONBody defines parameters for PostBbsAccountsCount.
+type PostBbsAccountsCountJSONBody = map[string]interface{}
+
+// PostBbsAccountsCountParams defines parameters for PostBbsAccountsCount.
+type PostBbsAccountsCountParams struct {
+	// The application ID to use to filter accounts
+	AppId *string `form:"app_id,omitempty" json:"app_id,omitempty"`
+
+	// The organization ID to use to filter accounts
+	OrgId *string `form:"org_id,omitempty" json:"org_id,omitempty"`
+}
 
 // PostBbsServiceAccountIdJSONBody defines parameters for PostBbsServiceAccountId.
 type PostBbsServiceAccountIdJSONBody = ServicesReqServiceAccountsParams
@@ -1634,6 +1670,9 @@ type GetSystemServiceAccountsParams struct {
 
 	// A comma-separated list of service account permissions to search for
 	Permissions *string `form:"permissions,omitempty" json:"permissions,omitempty"`
+
+	// A comma-separated list of service account scopes to search for
+	Scopes *string `form:"scopes,omitempty" json:"scopes,omitempty"`
 }
 
 // PostSystemServiceAccountsJSONBody defines parameters for PostSystemServiceAccounts.
@@ -1710,6 +1749,36 @@ type PostTpsAccessTokenJSONBody = ServicesReqServiceAccountsAccessToken
 
 // PostTpsAccessTokensJSONBody defines parameters for PostTpsAccessTokens.
 type PostTpsAccessTokensJSONBody = ServicesReqServiceAccountsAccessTokens
+
+// PostTpsAccountsJSONBody defines parameters for PostTpsAccounts.
+type PostTpsAccountsJSONBody = map[string]interface{}
+
+// PostTpsAccountsParams defines parameters for PostTpsAccounts.
+type PostTpsAccountsParams struct {
+	// The application ID to use to filter accounts
+	AppId *string `form:"app_id,omitempty" json:"app_id,omitempty"`
+
+	// The organization ID to use to filter accounts
+	OrgId *string `form:"org_id,omitempty" json:"org_id,omitempty"`
+
+	// The maximum number of accounts to return
+	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
+
+	// The index of the first account to return
+	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
+}
+
+// PostTpsAccountsCountJSONBody defines parameters for PostTpsAccountsCount.
+type PostTpsAccountsCountJSONBody = map[string]interface{}
+
+// PostTpsAccountsCountParams defines parameters for PostTpsAccountsCount.
+type PostTpsAccountsCountParams struct {
+	// The application ID to use to filter accounts
+	AppId *string `form:"app_id,omitempty" json:"app_id,omitempty"`
+
+	// The organization ID to use to filter accounts
+	OrgId *string `form:"org_id,omitempty" json:"org_id,omitempty"`
+}
 
 // PostTpsServiceAccountIdJSONBody defines parameters for PostTpsServiceAccountId.
 type PostTpsServiceAccountIdJSONBody = ServicesReqServiceAccountsParams
@@ -1800,6 +1869,12 @@ type PostBbsAccessTokenJSONRequestBody = PostBbsAccessTokenJSONBody
 
 // PostBbsAccessTokensJSONRequestBody defines body for PostBbsAccessTokens for application/json ContentType.
 type PostBbsAccessTokensJSONRequestBody = PostBbsAccessTokensJSONBody
+
+// PostBbsAccountsJSONRequestBody defines body for PostBbsAccounts for application/json ContentType.
+type PostBbsAccountsJSONRequestBody = PostBbsAccountsJSONBody
+
+// PostBbsAccountsCountJSONRequestBody defines body for PostBbsAccountsCount for application/json ContentType.
+type PostBbsAccountsCountJSONRequestBody = PostBbsAccountsCountJSONBody
 
 // PostBbsServiceAccountIdJSONRequestBody defines body for PostBbsServiceAccountId for application/json ContentType.
 type PostBbsServiceAccountIdJSONRequestBody = PostBbsServiceAccountIdJSONBody
@@ -1947,6 +2022,12 @@ type PostTpsAccessTokenJSONRequestBody = PostTpsAccessTokenJSONBody
 
 // PostTpsAccessTokensJSONRequestBody defines body for PostTpsAccessTokens for application/json ContentType.
 type PostTpsAccessTokensJSONRequestBody = PostTpsAccessTokensJSONBody
+
+// PostTpsAccountsJSONRequestBody defines body for PostTpsAccounts for application/json ContentType.
+type PostTpsAccountsJSONRequestBody = PostTpsAccountsJSONBody
+
+// PostTpsAccountsCountJSONRequestBody defines body for PostTpsAccountsCount for application/json ContentType.
+type PostTpsAccountsCountJSONRequestBody = PostTpsAccountsCountJSONBody
 
 // PostTpsServiceAccountIdJSONRequestBody defines body for PostTpsServiceAccountId for application/json ContentType.
 type PostTpsServiceAccountIdJSONRequestBody = PostTpsServiceAccountIdJSONBody